Refactor commonJs to ECMA
